Revert "[Data] Implement Operators for `union()` (#36242)"
This reverts commit d2073615b1cbedb148d792290442e75892e77456.
