+++ python/ray/data/_internal/execution/operators/actor_pool_map_operator.py
@@ -230,10 +230,10 @@ class ActorPoolMapOperator(MapOperator):
         # For either a completed task or ready worker, we try to dispatch queued tasks.
         self._dispatch_tasks()
 
-    def all_inputs_done(self):
+    def inputs_done(self):
         # Call base implementation to handle any leftover bundles. This may or may not
         # trigger task dispatch.
-        super().all_inputs_done()
+        super().inputs_done()
 
         # Mark inputs as done so future task dispatch will kill all inactive workers
         # once the bundle queue is exhausted.
