+++ python/ray/data/_internal/planner/plan_all_to_all_op.py
@@ -1,7 +1,5 @@
 from ray.data._internal.execution.interfaces import PhysicalOperator
-from ray.data._internal.execution.operators.base_physical_operator import (
-    AllToAllOperator,
-)
+from ray.data._internal.execution.operators.all_to_all_operator import AllToAllOperator
 from ray.data._internal.logical.operators.all_to_all_operator import (
     AbstractAllToAll,
     Aggregate,
