+++ python/ray/data/_internal/planner/planner.py
@@ -1,7 +1,6 @@
 from typing import Dict
 
 from ray.data._internal.execution.interfaces import PhysicalOperator
-from ray.data._internal.execution.operators.union_operator import UnionOperator
 from ray.data._internal.execution.operators.zip_operator import ZipOperator
 from ray.data._internal.logical.interfaces import (
     LogicalOperator,
@@ -12,7 +11,7 @@ from ray.data._internal.logical.operators.all_to_all_operator import AbstractAll
 from ray.data._internal.logical.operators.from_operators import AbstractFrom
 from ray.data._internal.logical.operators.input_data_operator import InputData
 from ray.data._internal.logical.operators.map_operator import AbstractUDFMap
-from ray.data._internal.logical.operators.n_ary_operator import Union, Zip
+from ray.data._internal.logical.operators.n_ary_operator import Zip
 from ray.data._internal.logical.operators.one_to_one_operator import Limit
 from ray.data._internal.logical.operators.read_operator import Read
 from ray.data._internal.logical.operators.write_operator import Write
@@ -67,9 +66,6 @@ class Planner:
         elif isinstance(logical_op, Zip):
             assert len(physical_children) == 2
             physical_op = ZipOperator(physical_children[0], physical_children[1])
-        elif isinstance(logical_op, Union):
-            assert len(physical_children) >= 2
-            physical_op = UnionOperator(*physical_children)
         elif isinstance(logical_op, Limit):
             assert len(physical_children) == 1
             physical_op = _plan_limit_op(logical_op, physical_children[0])
