+++ python/ray/data/dataset.py
@@ -57,9 +57,6 @@ from ray.data._internal.logical.operators.map_operator import (
     MapBatches,
     MapRows,
 )
-from ray.data._internal.logical.operators.n_ary_operator import (
-    Union as UnionLogicalOperator,
-)
 from ray.data._internal.logical.operators.n_ary_operator import Zip
 from ray.data._internal.logical.operators.one_to_one_operator import Limit
 from ray.data._internal.logical.operators.write_operator import Write
@@ -1618,25 +1615,14 @@ class Dataset:
         if has_nonlazy:
             blocks = []
             metadata = []
-            ops_to_union = []
-            for idx, bl in enumerate(bls):
+            for bl in bls:
                 if isinstance(bl, LazyBlockList):
                     bs, ms = bl._get_blocks_with_metadata()
                 else:
-                    assert isinstance(bl, BlockList), type(bl)
                     bs, ms = bl._blocks, bl._metadata
-                op_logical_plan = getattr(datasets[idx]._plan, "_logical_plan", None)
-                if isinstance(op_logical_plan, LogicalPlan):
-                    ops_to_union.append(op_logical_plan.dag)
-                else:
-                    ops_to_union.append(None)
                 blocks.extend(bs)
                 metadata.extend(ms)
             blocklist = BlockList(blocks, metadata, owned_by_consumer=owned_by_consumer)
-
-            logical_plan = None
-            if all(ops_to_union):
-                logical_plan = LogicalPlan(UnionLogicalOperator(*ops_to_union))
         else:
             tasks: List[ReadTask] = []
             block_partition_refs: List[ObjectRef[BlockPartition]] = []
@@ -1664,16 +1650,6 @@ class Dataset:
                 owned_by_consumer=owned_by_consumer,
             )
 
-            logical_plan = self._logical_plan
-            logical_plans = [
-                getattr(union_ds, "_logical_plan", None) for union_ds in datasets
-            ]
-            if all(logical_plans):
-                op = UnionLogicalOperator(
-                    *[plan.dag for plan in logical_plans],
-                )
-                logical_plan = LogicalPlan(op)
-
         epochs = [ds._get_epoch() for ds in datasets]
         max_epoch = max(*epochs)
         if len(set(epochs)) > 1:
@@ -1693,7 +1669,6 @@ class Dataset:
             ExecutionPlan(blocklist, stats, run_by_consumer=owned_by_consumer),
             max_epoch,
             self._lazy,
-            logical_plan,
         )
 
     def groupby(self, key: Optional[str]) -> "GroupedData":
