+++ python/ray/data/tests/test_execution_optimizer.py
@@ -6,14 +6,10 @@ import pandas as pd
 import pytest
 
 import ray
-from ray.data._internal.execution.interfaces import ExecutionOptions
 from ray.data._internal.execution.legacy_compat import _blocks_to_input_buffer
-from ray.data._internal.execution.operators.base_physical_operator import (
-    AllToAllOperator,
-)
+from ray.data._internal.execution.operators.all_to_all_operator import AllToAllOperator
 from ray.data._internal.execution.operators.input_data_buffer import InputDataBuffer
 from ray.data._internal.execution.operators.map_operator import MapOperator
-from ray.data._internal.execution.operators.union_operator import UnionOperator
 from ray.data._internal.execution.operators.zip_operator import ZipOperator
 from ray.data._internal.logical.interfaces import LogicalPlan
 from ray.data._internal.logical.operators.all_to_all_operator import (
@@ -34,7 +30,7 @@ from ray.data._internal.logical.operators.map_operator import (
     MapBatches,
     MapRows,
 )
-from ray.data._internal.logical.operators.n_ary_operator import Union, Zip
+from ray.data._internal.logical.operators.n_ary_operator import Zip
 from ray.data._internal.logical.operators.read_operator import Read
 from ray.data._internal.logical.operators.write_operator import Write
 from ray.data._internal.logical.optimizers import PhysicalOptimizer
@@ -46,8 +42,6 @@ from ray.data._internal.logical.util import (
 from ray.data._internal.planner.planner import Planner
 from ray.data._internal.stats import DatasetStats
 from ray.data.aggregate import Count
-from ray.data.datasource.datasource import RangeDatasource
-from ray.data.datasource.json_datasource import JSONDatasource
 from ray.data.datasource.parquet_datasource import ParquetDatasource
 from ray.data.tests.conftest import *  # noqa
 from ray.data.tests.util import column_udf, extract_values, named_values
@@ -378,74 +372,6 @@ def test_repartition_e2e(
     assert ds.sum() == sum(range(10))
     assert ds._block_num_rows() == [10], ds._block_num_rows()
     _check_repartition_usage_and_stats(ds)
-
-
-@pytest.mark.parametrize("preserve_order", (True, False))
-def test_union_operator(ray_start_regular_shared, enable_optimizer, preserve_order):
-    planner = Planner()
-    read_parquet_op = Read(ParquetDatasource(), [])
-    read_range_op = Read(RangeDatasource(), [])
-    read_json_op = Read(JSONDatasource(), [])
-    union_op = Union(
-        read_parquet_op,
-        read_range_op,
-        read_json_op,
-    )
-    plan = LogicalPlan(union_op)
-    physical_op = planner.plan(plan).dag
-
-    assert union_op.name == "Union"
-    assert isinstance(physical_op, UnionOperator)
-    assert len(physical_op.input_dependencies) == 3
-    for input_op in physical_op.input_dependencies:
-        assert isinstance(input_op, MapOperator)
-
-
-@pytest.mark.parametrize("preserve_order", (True, False))
-def test_union_e2e(ray_start_regular_shared, enable_optimizer, preserve_order):
-    execution_options = ExecutionOptions(preserve_order=preserve_order)
-    ctx = ray.data.DataContext.get_current()
-    ctx.execution_options = execution_options
-
-    ds = ray.data.range(20, parallelism=10)
-
-    # Test lazy union.
-    ds = ds.union(ds, ds, ds, ds)
-    assert ds.num_blocks() == 50
-    assert ds.count() == 100
-    assert ds.sum() == 950
-    _check_usage_record(["ReadRange", "Union"])
-    ds_result = [{"id": i} for i in range(20)] * 5
-    if preserve_order:
-        assert ds.take_all() == ds_result
-
-    ds = ds.union(ds)
-    assert ds.count() == 200
-    assert ds.sum() == (950 * 2)
-    _check_usage_record(["ReadRange", "Union"])
-    if preserve_order:
-        assert ds.take_all() == ds_result * 2
-
-    # Test materialized union.
-    ds2 = ray.data.from_items([{"id": i} for i in range(1, 5 + 1)])
-    assert ds2.count() == 5
-    assert ds2.sum() == 15
-    _check_usage_record(["FromItems"])
-
-    ds2 = ds2.union(ds2)
-    assert ds2.count() == 10
-    assert ds2.sum() == 30
-    _check_usage_record(["FromItems", "Union"])
-    ds2_result = ([{"id": i} for i in range(1, 5 + 1)]) * 2
-    if preserve_order:
-        assert ds2.take_all() == ds2_result
-
-    ds2 = ds2.union(ds)
-    assert ds2.count() == 210
-    assert ds2.sum() == (950 * 2 + 30)
-    _check_usage_record(["FromItems", "Union"])
-    if preserve_order:
-        assert ds2.take_all() == (ds2_result + ds_result * 2)
 
 
 def test_read_map_batches_operator_fusion(ray_start_regular_shared, enable_optimizer):
