+++ python/ray/data/tests/test_streaming_executor.py
@@ -25,7 +25,6 @@ from ray.data._internal.execution.streaming_executor_state import (
     build_streaming_topology,
     process_completed_tasks,
     select_operator_to_run,
-    update_operator_states,
 )
 from ray.data._internal.execution.util import make_ref_bundles
 from ray.data.tests.conftest import *  # noqa
@@ -92,7 +91,6 @@ def test_process_completed_tasks():
     # Test processing output bundles.
     assert len(topo[o1].outqueue) == 0, topo
     process_completed_tasks(topo)
-    update_operator_states(topo)
     assert len(topo[o1].outqueue) == 20, topo
 
     # Test processing completed work items.
@@ -100,32 +98,29 @@ def test_process_completed_tasks():
     done_ref = ray.put("done")
     o2.get_work_refs = MagicMock(return_value=[sleep_ref, done_ref])
     o2.notify_work_completed = MagicMock()
-    o2.all_inputs_done = MagicMock()
+    o2.inputs_done = MagicMock()
     o1.all_dependents_complete = MagicMock()
     process_completed_tasks(topo)
-    update_operator_states(topo)
     o2.notify_work_completed.assert_called_once_with(done_ref)
-    o2.all_inputs_done.assert_not_called()
+    o2.inputs_done.assert_not_called()
     o1.all_dependents_complete.assert_not_called()
 
     # Test input finalization.
     o2.get_work_refs = MagicMock(return_value=[done_ref])
     o2.notify_work_completed = MagicMock()
-    o2.all_inputs_done = MagicMock()
+    o2.inputs_done = MagicMock()
     o1.all_dependents_complete = MagicMock()
     o1.completed = MagicMock(return_value=True)
     topo[o1].outqueue.clear()
     process_completed_tasks(topo)
-    update_operator_states(topo)
     o2.notify_work_completed.assert_called_once_with(done_ref)
-    o2.all_inputs_done.assert_called_once()
+    o2.inputs_done.assert_called_once()
     o1.all_dependents_complete.assert_not_called()
 
     # Test dependents completed.
     o2.need_more_inputs = MagicMock(return_value=False)
     o1.all_dependents_complete = MagicMock()
     process_completed_tasks(topo)
-    update_operator_states(topo)
     o1.all_dependents_complete.assert_called_once()
