+++ python/ray/data/tests/test_streaming_integration.py
@@ -15,9 +15,7 @@ from ray.data._internal.execution.interfaces import (
     ExecutionResources,
     RefBundle,
 )
-from ray.data._internal.execution.operators.base_physical_operator import (
-    AllToAllOperator,
-)
+from ray.data._internal.execution.operators.all_to_all_operator import AllToAllOperator
 from ray.data._internal.execution.operators.input_data_buffer import InputDataBuffer
 from ray.data._internal.execution.operators.map_operator import MapOperator
 from ray.data._internal.execution.operators.output_splitter import OutputSplitter
